+++ spring/TFG_inma_spring/src/main/java/com/tfginma/app/valoracion/domain/Valoracion.java
@@ -30,7 +30,7 @@ public class Valoracion {
 	private Long idValoracion;
 	
 	
-	private Integer valor;
+	private Float valor;
 	
 	@OneToMany(mappedBy="valoracion")
 	@Cascade(CascadeType.ALL)
